# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: La Liga Pro Analytics — live data backed by ESPN public sports APIs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Team detail with squad and form
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamDetail" }
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches filtered by status and/or gameweek (live from ESPN)
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [live, upcoming, finished, all]
        - in: query
          name: gameweek
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MatchSummary" }
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Full match detail (stats, lineups, momentum, referee)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Match detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchDetail" }
...
  /predictions/{matchId}:
    get:
      operationId: getPrediction
      tags: [predictions]
      summary: Detailed match prediction (Poisson, form, H2H, absences, market value, props)
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Prediction
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PredictionDetail" }
  /predictions/{matchId}/players:
    get:
      operationId: getPlayerProps
      tags: [predictions]
      summary: Per-player prop predictions for a match (anytime scorer, assist, goal contribution)
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Per-player predictions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerPropsResponse" }
  /predictions/{matchId}/lineups:
    get:
      operationId: getProbableLineups
      tags: [predictions]
      summary: Probable starting XI and bench for both sides
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Probable lineups
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProbableLineupsResponse" }
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: La Liga players (filterable by team / position)
      parameters:
        - in: query
          name: teamId
          required: false
          schema: { type: integer }
        - in: query
          name: position
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Players
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Player" }
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Player profile and season metrics
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Player detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerDetail" }
...
components:
  schemas:
...
    DataSource:
      type: object
      description: Provenance metadata for live data fetched from ESPN.
      properties:
        provider: { type: string, description: "Upstream data provider (e.g. ESPN, DraftKings)" }
        fetchedAt: { type: string, format: date-time }
        cacheTtlSeconds: { type: integer }
      required: [provider, fetchedAt]
    Team:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        shortName: { type: string }
        abbreviation: { type: string }
        city: { type: string }
        founded: { type: integer, nullable: true }
        stadium: { type: string }
        primaryColor: { type: string }
        secondaryColor: { type: string }
        crestUrl: { type: string }
        manager: { type: string }
        formation: { type: string }
      required: [id, name, shortName, abbreviation, city, stadium, primaryColor, secondaryColor, crestUrl, manager, formation]
    TeamDetail:
      type: object
      properties:
        team: { $ref: "#/components/schemas/Team" }
        position: { type: integer }
        points: { type: integer }
        played: { type: integer }
        wins: { type: integer }
        draws: { type: integer }
        losses: { type: integer }
        goalsFor: { type: integer }
        goalsAgainst: { type: integer }
        cleanSheets: { type: integer }
        form:
          type: array
          items: { type: string, enum: [W, D, L] }
        squad:
          type: array
          items: { $ref: "#/components/schemas/Player" }
        upcomingMatches:
          type: array
          items: { $ref: "#/components/schemas/MatchSummary" }
        dataSource: { $ref: "#/components/schemas/DataSource" }
      required: [team, position, points, played, wins, draws, losses, goalsFor, goalsAgainst, cleanSheets, form, squad, upcomingMatches]
...
    MatchSummary:
      type: object
      properties:
        id: { type: integer }
        gameweek: { type: integer }
        kickoff: { type: string, format: date-time }
        status: { type: string, enum: [live, upcoming, finished] }
        statusDetail: { type: string }
        minute: { type: integer, nullable: true }
        homeTeam: { $ref: "#/components/schemas/Team" }
        awayTeam: { $ref: "#/components/schemas/Team" }
        homeScore: { type: integer, nullable: true }
        awayScore: { type: integer, nullable: true }
        venue: { type: string }
        referee: { type: string }
      required: [id, gameweek, kickoff, status, statusDetail, homeTeam, awayTeam, venue, referee]
    MatchDetail:
      type: object
      properties:
        match: { $ref: "#/components/schemas/MatchSummary" }
        stats: { $ref: "#/components/schemas/MatchStats" }
        homeLineup: { $ref: "#/components/schemas/Lineup" }
        awayLineup: { $ref: "#/components/schemas/Lineup" }
        momentum:
          type: array
          items: { $ref: "#/components/schemas/MomentumPoint" }
        events:
          type: array
          items: { $ref: "#/components/schemas/MatchEvent" }
        refereeStats: { $ref: "#/components/schemas/RefereeStats" }
        dataSource: { $ref: "#/components/schemas/DataSource" }
      required: [match, stats, homeLineup, awayLineup, momentum, events, refereeStats]
    MatchStats:
      type: object
      properties:
        homePossession: { type: number }
        awayPossession: { type: number }
        homeShots: { type: number }
        awayShots: { type: number }
        homeShotsOnTarget: { type: number }
        awayShotsOnTarget: { type: number }
        homeShotsOffTarget: { type: number }
        awayShotsOffTarget: { type: number }
        homeShotsBlocked: { type: number }
        awayShotsBlocked: { type: number }
        homeCorners: { type: number }
        awayCorners: { type: number }
        homeOffsides: { type: number }
        awayOffsides: { type: number }
        homeFouls: { type: number }
        awayFouls: { type: number }
        homeYellow: { type: number }
        awayYellow: { type: number }
        homeRed: { type: number }
        awayRed: { type: number }
        homeSaves: { type: number }
        awaySaves: { type: number }
        homeXG: { type: number }
        awayXG: { type: number }
        homePassAccuracy: { type: number }
        awayPassAccuracy: { type: number }
      required: [homePossession, awayPossession, homeShots, awayShots, homeShotsOnTarget, awayShotsOnTarget, homeShotsOffTarget, awayShotsOffTarget, homeShotsBlocked, awayShotsBlocked, homeCorners, awayCorners, homeOffsides, awayOffsides, homeFouls, awayFouls, homeYellow, awayYellow, homeRed, awayRed, homeSaves, awaySaves, homeXG, awayXG, homePassAccuracy, awayPassAccuracy]
    Lineup:
      type: object
      properties:
        formation: { type: string }
        source: { type: string, enum: [official, predicted] }
        starting:
          type: array
          items: { $ref: "#/components/schemas/LineupPlayer" }
        bench:
          type: array
          items: { $ref: "#/components/schemas/LineupPlayer" }
      required: [formation, source, starting, bench]
...
    MomentumPoint:
      type: object
      properties:
        minute: { type: integer }
        homeIntensity: { type: number }
        awayIntensity: { type: number }
      required: [minute, homeIntensity, awayIntensity]
    MatchEvent:
      type: object
      properties:
        minute: { type: integer }
        type: { type: string, enum: [goal, yellow, red, sub, var] }
        teamSide: { type: string, enum: [home, away] }
        playerName: { type: string }
        detail: { type: string, nullable: true }
      required: [minute, type, teamSide, playerName]
    RefereeStats:
      type: object
      properties:
        name: { type: string }
        matches: { type: integer }
        avgYellow: { type: number }
        avgRed: { type: number }
        avgFouls: { type: number }
        penaltiesGiven: { type: integer }
      required: [name, matches, avgYellow, avgRed, avgFouls, penaltiesGiven]
    Player:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        teamId: { type: integer }
        teamShortName: { type: string }
        position: { type: string, enum: [GK, DEF, MID, FWD] }
        positionLabel: { type: string }
        shirtNumber: { type: integer, nullable: true }
        nationality: { type: string }
        age: { type: integer, nullable: true }
        headshotUrl: { type: string, nullable: true }
        appearances: { type: number }
        goals: { type: number }
        assists: { type: number }
        keyPasses: { type: number }
        bigChancesCreated: { type: number }
        shots: { type: number }
        shotsOnTarget: { type: number }
        xG: { type: number }
        rating: { type: number }
        injured: { type: boolean }
        injuryStatus: { type: string, nullable: true }
      required: [id, name, teamId, teamShortName, position, positionLabel, nationality, appearances, goals, assists, keyPasses, bigChancesCreated, shots, shotsOnTarget, xG, rating, injured]
    PlayerDetail:
      type: object
      properties:
        player: { $ref: "#/components/schemas/Player" }
        radar:
          type: array
          items:
            type: object
            properties:
              metric: { type: string }
              value: { type: number }
              max: { type: number }
            required: [metric, value, max]
        recentForm:
          type: array
          items:
            type: object
            properties:
              matchLabel: { type: string }
              rating: { type: number }
              goals: { type: integer }
              assists: { type: integer }
            required: [matchLabel, rating, goals, assists]
      required: [player, radar, recentForm]
    InjuryReport:
      type: object
      properties:
        id: { type: integer }
        playerId: { type: integer }
        playerName: { type: string }
        teamId: { type: integer }
        teamName: { type: string }
        teamShortName: { type: string }
        position: { type: string, enum: [GK, DEF, MID, FWD] }
        type: { type: string, enum: [injury, suspension, doubtful] }
        status: { type: string }
        description: { type: string }
        severity: { type: string, enum: [low, medium, high] }
        expectedReturn: { type: string, nullable: true }
        impactScore: { type: number }
      required: [id, playerId, playerName, teamId, teamName, teamShortName, position, type, status, description, severity, impactScore]
    Prediction:
      type: object
      properties:
        matchId: { type: integer }
        kickoff: { type: string, format: date-time }
        homeTeam: { $ref: "#/components/schemas/Team" }
        awayTeam: { $ref: "#/components/schemas/Team" }
        homeWinProb: { type: number }
        drawProb: { type: number }
        awayWinProb: { type: number }
        expectedHomeGoals: { type: number }
        expectedAwayGoals: { type: number }
        bttsProb: { type: number }
        over25Prob: { type: number }
        under25Prob: { type: number }
        cleanSheetHome: { type: number }
        cleanSheetAway: { type: number }
        confidence: { type: number }
        recommendation: { type: string }
        source: { type: string, enum: [bookmaker, team-rates, blended] }
        bookmaker: { type: string, nullable: true }
        oddsLastUpdate: { type: string, format: date-time }
      required: [matchId, kickoff, homeTeam, awayTeam, homeWinProb, drawProb, awayWinProb, expectedHomeGoals, expectedAwayGoals, bttsProb, over25Prob, under25Prob, cleanSheetHome, cleanSheetAway, confidence, recommendation, source, oddsLastUpdate]
    PredictionDetail:
      type: object
      properties:
        base: { $ref: "#/components/schemas/Prediction" }
        scoreMatrix:
          type: array
          items:
            type: object
            properties:
              homeGoals: { type: integer }
              awayGoals: { type: integer }
              probability: { type: number }
            required: [homeGoals, awayGoals, probability]
        topScorelines:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              probability: { type: number }
            required: [label, probability]
        formFactor:
          type: object
          properties:
            homeFormScore: { type: number }
            awayFormScore: { type: number }
            homeLast5:
              type: array
              items: { type: string, enum: [W, D, L] }
            awayLast5:
              type: array
              items: { type: string, enum: [W, D, L] }
          required: [homeFormScore, awayFormScore, homeLast5, awayLast5]
        h2h: { $ref: "#/components/schemas/H2HSummary" }
        absenceImpact:
          type: object
          properties:
            homeImpact: { type: number }
            awayImpact: { type: number }
            homeMissing:
              type: array
              items: { $ref: "#/components/schemas/InjuryReport" }
            awayMissing:
              type: array
              items: { $ref: "#/components/schemas/InjuryReport" }
          required: [homeImpact, awayImpact, homeMissing, awayMissing]
        marketOdds: { $ref: "#/components/schemas/MarketView" }
        props: { $ref: "#/components/schemas/MatchProps" }
        playerProps: { $ref: "#/components/schemas/PlayerPropsResponse" }
        probableLineups: { $ref: "#/components/schemas/ProbableLineupsResponse" }
        dataSource: { $ref: "#/components/schemas/DataSource" }
      required: [base, scoreMatrix, topScorelines, formFactor, h2h, absenceImpact, marketOdds, props, playerProps, probableLineups]
    MarketView:
      type: object
      properties:
        homeOdds: { type: number }
        drawOdds: { type: number }
        awayOdds: { type: number }
        valueHome: { type: number }
        valueDraw: { type: number }
        valueAway: { type: number }
        bestPick: { type: string }
        source: { type: string, enum: [bookmaker, model] }
        bookmaker: { type: string, nullable: true }
        overUnderLine: { type: number, nullable: true }
        overOdds: { type: number, nullable: true }
        underOdds: { type: number, nullable: true }
      required: [homeOdds, drawOdds, awayOdds, valueHome, valueDraw, valueAway, bestPick, source]
    MatchProps:
      type: object
      properties:
        totalGoals:
          type: object
          properties:
            line: { type: number }
            expected: { type: number }
            overProb: { type: number }
            underProb: { type: number }
          required: [line, expected, overProb, underProb]
        totalCorners:
          type: object
          properties:
            line: { type: number }
            expected: { type: number }
          required: [line, expected]
        totalCards:
          type: object
          properties:
            line: { type: number }
            expected: { type: number }
          required: [line, expected]
        totalOffsides:
          type: object
          properties:
            line: { type: number }
            expected: { type: number }
          required: [line, expected]
        totalFouls:
          type: object
          properties:
            line: { type: number }
            expected: { type: number }
          required: [line, expected]
        bothTeamsToScore:
          type: object
          properties:
            yesProb: { type: number }
            noProb: { type: number }
          required: [yesProb, noProb]
        exactScore:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              probability: { type: number }
            required: [label, probability]
      required: [totalGoals, totalCorners, totalCards, totalOffsides, totalFouls, bothTeamsToScore, exactScore]
...
    PlayerPropsResponse:
      type: object
      properties:
        matchId: { type: integer }
        home:
          type: array
          items: { $ref: "#/components/schemas/PlayerProp" }
        away:
          type: array
          items: { $ref: "#/components/schemas/PlayerProp" }
        dataSource: { $ref: "#/components/schemas/DataSource" }
      required: [matchId, home, away]
...
    ProbableLineupsResponse:
      type: object
      properties:
        matchId: { type: integer }
        home: { $ref: "#/components/schemas/ProbableLineup" }
        away: { $ref: "#/components/schemas/ProbableLineup" }
        dataSource: { $ref: "#/components/schemas/DataSource" }
      required: [matchId, home, away]
    H2HSummary:
      type: object
      properties:
        homeTeamId: { type: integer }
        awayTeamId: { type: integer }
        totalMatches: { type: integer }
        homeWins: { type: integer }
        draws: { type: integer }
        awayWins: { type: integer }
        avgGoals: { type: number }
        recentMeetings:
          type: array
          items:
            type: object
            properties:
              date: { type: string }
              competition: { type: string }
              homeTeam: { type: string }
              awayTeam: { type: string }
              homeScore: { type: integer }
              awayScore: { type: integer }
            required: [date, competition, homeTeam, awayTeam, homeScore, awayScore]
      required: [homeTeamId, awayTeamId, totalMatches, homeWins, draws, awayWins, avgGoals, recentMeetings]
...